Improve esm support (#263)
## ESM support change
package/src/languageService/kustoLanguageService.ts now imports
"language-server" which the build script replaces with either esm
imports or `importScripts` depending on the module target

## Other approaches attempted

* **Different entry points for esm and amd bundles**: Rollup kept
putting `importScripts` at the bottom of the amd entry point, causing
`Kusto is undefined` errors. We need the language server files to have
been parsed `before` our language server files.
* **Using rollup's `replace` plugin**: Replace a comment like `///
LANGUAGE_SERVER_IMPORT` with the appropriate import given the build
target. The comment never got replaced and I couldn't figure out why 😢.

* **Make the modules containing the imports actual files, not virtual**:
and then alias to the appropriate one depending on the build target. If
we encounter problems with the current approach this is might be worth
trying.

* **Use amd imports instead of `requireScripts()`**: I tried using the
esm imports in the code, and then aliasing them in rollup to be the
right amd imports. Unfortunately, the monaco editor doesn't pass loader
config to it's workers, so my first attempt at this didn't work. It's
possible I might have just gotten the relative paths wrong, and it could
work? I'm not sure it would have any advantages over the current
approach if it did. Maybe simpler rollup config?

## Other notable changes
* Removed references to `Kusto` global in languageFeatures.ts and
kustoMonarchLanguageDefinition.ts so we can avoid loading the language
server on the main thread. This should both improve performance, and
simplify integration.
* Added a script for amd package consumers to copy files into their
`/vs` folder
* Added a `monaco-editor` import to monaco.contribution to ensure
"monaco" global is available before file is parsed. Replaced with
`vs/editor/editor.main` for amd builds. This removes the need to specify
this dependency in the loaders "shim" configuration
* Added a new minimal amd example
* Replaced `monaco-editor-core` imports with `monaco-editor`. My
understanding is that `monaco-editor-core` makes sense if we're bundling
parts of the monaco editor into our built files. We are not, so
depending on it just creates differences between us and our package
consumers.
* `importScripts` now called once with a list of files, that than being
called multiple times so the scripts are loaded in parallel