Added Parcel sample and fixed esm output issue (#264)
* Added Parcel.js sample
* Fixed kusto language server import order issue in esm output
* Imports are now for the same files as amd version, and in the same
order
* Updated README.md instructions to reflect latest changes
* Turned off yarn "nmHoistingLimits" feature. New build scripts use
`require.resolve()` to find dependencies so we don't need this anymore,
and it causes issues when using `yarn link` to run samples because we'll
have multiple copies of `monaco-kusto` in the tree
* Fixed worker consuming monaco global api
  * package/src/kustoWorker.ts
* I'm not actually sure if this code was not working, if this means
we're importing an extra copy of monaco at runtime, or something else
* Change was made while I was looking into updating us to not use the
global api. Didn't do this because the patch would have been even
bigger, and I was running into typing issues.
* Added patch to monaco-editor to fix Parcel compatibility issue
* PR to upstream this patch here:
https://github.com/microsoft/vscode/pull/178179
* Release pipeline now copies README.md into the package folder
* Updated package json with esm entry points, and removed "main" entry
point because we don't have a common.js entry point.