Add types to esm module interface (#275)
* Added types to @kusto/monaco-kusto exports
* Consumers of the esm files no longer need to import the global types
to get correct typings
* Global interface now references this one, so we don't need to keep
both in sync manually
* Stopped exporting `LanguageServiceDefaultsImpl` from
`@kusto/monaco-kusto` because I don't think we ever intended to export
this? We're already exporting the interface it implements.
* Updated webpack amd example to include types so we have a place to
test global api type changes
* CI now fails if there's a type error in either of the typescript
examples (amd-webpack-react or parcel)
   * Added Yarn workspace tools to do this
* Added `getCurrentCommandRange` to @kusto/monaco-kusto exports so we
have a type-safe way to call it when we're not using the global api.
This is because it's not possible to augment the non-global editor
interface. This way is tidier anyhow.
* Started publishing the global types in a new location:
`@kusto/monao-kusto/globalApi`. This makes it more clear that you don't
need to import this if you're not using the global api.
* Fixed type errors in source code that we're uncovered when I updated
it to not rely on the types in monaco.d.ts
* Updated typescript samples tsconfig files to extend @tsonfig/*
packages when possible