Added types back to global api (#284)
When I moved the types to /src out of the global api I didn't test it
well enough and the _types_ types we're re-exported. There isn't a good
way to re-export these, so I had to do each of them by hand.

I've also added some references to them to the amd typescript sample to
catch any future regression.