Fix AMD worker language support imports (#288)
This PR adds a fix for the paths used to import additional scripts using
a web worker. When a web worker is used with the current relative paths
the imports will fail due to CORS, making the library unusable.

This change fixes the issue by allowing the base URL of the Monaco
environment to be used (if defined) to provide a complete absolute path,
allowing the scripts to be loaded from the same origin the worker is
running from. This then allows the library to be served via CDN.

---------

Co-authored-by: Maxwell Burson <maburson@microsoft.com>