Add `maximumDepthExceeded` result kind to getReferencedGlobalParams (#521)
getReferencedGlobalParams can fail. Currently, it does so silently. PR
updates it's returned values to include an error so we can let users
know when it's failing.

I considered making diagnostics and parameters 2 properties instead of
using a union, but, it seemed unlikely callers would know what to do
with the raw list, and as far as I know, this is the only one they might
care about.

Bug: https://msazure.visualstudio.com/One/_workitems/edit/29324049